chore(setup): moved task into gulpfile, renamed first install to "setup"
