+++ package.json
@@ -25,8 +25,7 @@
     "postinstall": "install-app-deps",
     "pack": "build --dir",
     "dist": "build",
-    "first-install": "node install && npm run rebuild",
-    "rebuild": "node rebuild",
+    "setup": "gulp first-install && gulp rebuild-native",
     "test-build": "gulp rename-index && build && gulp rename-index-back"
   },
   "directories": {
